fix: properly set tags when creating final Transaction
When creating the final Transaction on the
`userinput.TransactionFromData` we were not passing the `Tags` along,
but instead saving it as `Comment`. This worked well for printing but
it also meant that the transaction we were saving into the state did
not have Tags. This was causing a missbehavior of the tags input,
which would not have the tag of the latest transaction.

This commit:

1. Properly handle `Tags` on TransactionFromData
2. Properly handle `Tags` on the printer